fixed knex run migration cli commands working
