+++ knexfile.js
@@ -1,4 +1,4 @@
-// import config from './config/index.js';
+import config from './config/index.js';
 
 export const generateKnexConfig = (conf) => {
   return {
@@ -16,24 +16,18 @@ export const generateKnexConfig = (conf) => {
       max: conf.pool_max || 10,
     },
     migrations: {
-      tableName: 'knex_migrations',
+      tableName: '__knex_migrations',
       directory: './database/migrations',
       stub: './database/migrations/stub',
-      loadExtensions: ['.js']
     },
     seeds: {
       directory: './database/seeds',
     },
   };
 };
+const keys = Object.keys(config.DB);
+const autoConf = config.DB[keys[0]];
 
-// const keys = Object.keys(config.DB);
-// const autoConf = config.DB[keys[0]];
-
-
-export default async () => {
-  let config = await import('./config/index.js');
-  return {
-    [config.NODE_ENV]: await generateKnexConfig(autoConf),
-  };
+export default {
+  [config.NODE_ENV]:  generateKnexConfig(autoConf),
 }
