+++ lib/jsonLoader.js
@@ -5,4 +5,4 @@ import path from 'path';
 export const jsonLoader = async (filename) => JSON.parse(await readFile(path.resolve(filename)));
 export const jsonLoaderSync = (filename) => JSON.parse(readFileSync(path.resolve(filename)));
 
-export default await { jsonLoader, jsonLoaderSync };
+export default{ jsonLoader, jsonLoaderSync };
